added webpack, babel and vue-loader
