+++ package.json
@@ -4,8 +4,8 @@
   "description": "A adiciona funcionalidades e melhora a usabilidade do LegalOne. Esta é uma ferramenta open-source desenvolvida pela comunidade e não possui ligação com a funcionalidade com a Thomson Reuters</b>.",
   "main": "main.entry.js",
   "scripts": {
-    "build": "webpack",
-    "run": "webpack-dev-server --progress --colors"
+    "watch": "webpack --watch",
+    "build": "cross-env NODE_ENV=production webpack --progress"
   },
   "repository": {
     "type": "git",
@@ -22,6 +22,18 @@
     "vue": "^1.0.26"
   },
   "devDependencies": {
+    "babel-core": "^6.14.0",
+    "babel-loader": "^6.2.5",
+    "babel-plugin-transform-runtime": "^6.12.0",
+    "babel-plugin-uglify": "^1.0.2",
+    "babel-preset-es2015": "^6.14.0",
+    "babel-runtime": "^6.11.6",
+    "cross-env": "^2.0.0",
+    "css-loader": "^0.24.0",
+    "vue-hot-reload-api": "^2.0.6",
+    "vue-html-loader": "^1.2.3",
+    "vue-loader": "^8.5.2",
+    "vue-style-loader": "^1.0.0",
     "webpack": "^1.13.2",
     "webpack-dev-server": "^1.15.0"
   }
