+++ scripts/main.entry.js
@@ -4,6 +4,8 @@ var menu = require("./menu/menu")();
 var clone = require("./lib/clone");
 var searchUtils = require("./lib/searchUtils");
 
+var Vnav = require('./components/menu.vue');
+
 $(".main-bar")
     .before(`<div class='side-bar side-bar-left'>
                 <div id='vnav-container'>
@@ -16,33 +18,11 @@ $(".main-bar")
                 </div>
             </div>`);
 
-Vue.component('vnav', {
-    template: `<div class='vnav-children-wrapper'>
-                <div v-for="menuItem in menu" class='vnav-item-wrapper'>
-                    <div class='vnav-item' data-val-displayNameFull='{{menuItem.displayNameFull}}'>
-                        <i v-show="menuItem.icon" class="vnav-icon fa fa-{{menuItem.icon}}" ></i>
-                        <span class="vnav-item-text with-icon" >
-                            <a v-bind:href="menuItem.url" class="vnav-link" >{{ menuItem.displayName }}</a>
-                        </span>
-                        <span v-if="menuItem.menu && menuItem.menu.length" class="plus-wrapper" v-on:click="toggle(menuItem)" >
-                            <i class="vnav-icon fa" v-bind:class="menuItem.collapsed ? 'fa-plus' : 'fa-minus'" ></i>
-                        </span>
-                    </div>
-                    <vnav v-if="!menuItem.collapsed" :menu="menuItem.menu"/>
-                </div>
-            </div>`,
-    props: {
-        menu: Array
-    },
-    methods: {
-        toggle: function(menuItem) {            
-            menuItem.collapsed = !menuItem.collapsed;
-        }
-    }
-});
-
 new Vue({
     el: '#vnav-container',
+    components: {
+        vnav: Vnav
+    },
     data: function() {
         var menuItems = clone(menu);
         this.updateCollapsedState(menuItems, true);
